+++ components/SSI/DIDDocument/Keys/styles.module.scss
@@ -79,3 +79,21 @@
 .card:hover .cardTitle {
   display: none;
 }
+
+.button {
+  align-items: center;
+  justify-content: center;
+  border-radius: 6px;
+  outline: 0;
+  background-color: #ffffff2e;
+  cursor: pointer;
+  padding: 0 0.6rem 0 0.8rem;
+  letter-spacing: 0.2rem;
+  max-width: fit-content;
+}
+
+.buttonText {
+  margin: 0;
+  font-family: "Source Sans Pro", sans-serif;
+  font-weight: 400;
+}
