+++ components/SSI/DIDDocument/Services/styles.module.scss
@@ -3,7 +3,7 @@
   padding: 15px;
   border-radius: 5px;
   margin: 10px 0;
-  width: 103%;
+  width: 300px;
 }
 
 .blockHead {
@@ -78,4 +78,22 @@
 
 .card:hover .cardTitle {
   display: none;
+}
+
+.button {
+  align-items: center;
+  justify-content: center;
+  border-radius: 6px;
+  outline: 0;
+  background-color: #ffffff2e;
+  cursor: pointer;
+  padding: 0 0.6rem 0 0.8rem;
+  letter-spacing: 0.2rem;
+  max-width: fit-content;
+}
+
+.buttonText {
+  margin: 0;
+  font-family: "Source Sans Pro", sans-serif;
+  font-weight: 400;
 }
