+++ pages/[username]/did/services.tsx
@@ -1,0 +1,17 @@
+import Layout from "../../../components/Layout";
+import { SSI } from "../../../components";
+import Services from '../../../components/SSI/DIDDocument/Services'
+
+function Header() {
+  return (
+    <>
+      <Layout>
+        <SSI>
+          <Services />
+        </SSI>
+      </Layout>
+    </>
+  );
+}
+
+export default Header;
